shared/components/stepper/stepper.component.spec.ts
import { CdkStepperModule } from '@angular/cdk/stepper';
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { MatIconModule } from '@angular/material/icon';
import { MatProgressBarModule } from '@angular/material/progress-bar';
import { StepperComponent } from './stepper.component';

describe('StepperComponent', () => {
  let component: StepperComponent;
  let fixture: ComponentFixture<StepperComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [StepperComponent],
      imports: [MatProgressBarModule, MatIconModule, CdkStepperModule],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(StepperComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('progress should have initial value', () => {
    expect(component.progress).toBe(1);
  });

  it('close should emit event', () => {
    const spy = spyOn(component.stepperClose, 'emit');
    component.close();
    expect(spy).toHaveBeenCalled();
  });
});
